requirements.txt
python-telegram-bot>=21.0
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
